hasura-rest-api/requirements.txt
asgiref==3.3.4
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
drf-firebase-auth==1.0.0
firebase-admin==4.5.3
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
Markdown==3.3.4
msgpack==1.0.2
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
uritemplate==3.0.1
urllib3==1.26.5
